requirements.txt
aioredis==1.3.1
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
cheroot==8.6.0
CherryPy==18.6.1
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
flower==1.0.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
hiredis==2.0.0
humanize==4.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
kombu==5.2.4
more-itertools==8.12.0
msgpack==1.0.3
packaging==21.3
portend==3.1.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2021.3
redis==4.2.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tempora==5.0.1
tornado==6.1
Twisted==22.2.0
txaio==22.2.1
typing-extensions==4.1.1
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.2
wrapt==1.14.0
zc.lockfile==2.0
zope.event==4.5.0
zope.interface==5.4.0
